[ZCash] Add v6 tx format serializer.
