+++ components/brave_wallet/browser/BUILD.gn
@@ -298,6 +298,10 @@ static_library("browser") {
     "zcash/zcash_resolve_transaction_status_task.h",
     "zcash/zcash_rpc.cc",
     "zcash/zcash_rpc.h",
+    "zcash/v5_zcash_serializer.cc",
+    "zcash/v5_zcash_serializer.h",
+    "zcash/v6_zcash_serializer.cc",
+    "zcash/v6_zcash_serializer.h",
     "zcash/zcash_serializer.cc",
     "zcash/zcash_serializer.h",
     "zcash/zcash_transaction.cc",
@@ -818,6 +822,8 @@ source_set("unit_tests") {
     "zcash/zcash_grpc_utils_unittest.cc",
     "zcash/zcash_keyring_unittest.cc",
     "zcash/zcash_resolve_transaction_status_task_unittest.cc",
+    "zcash/v5_zcash_serializer_unittest.cc",
+    "zcash/v6_zcash_serializer_unittest.cc",
     "zcash/zcash_serializer_unittest.cc",
     "zcash/zcash_transaction_unittest.cc",
     "zcash/zcash_transaction_utils_unittest.cc",
