+++ components/brave_wallet/browser/zcash/zcash_serializer.cc
@@ -14,7 +14,8 @@
 #include "base/check_op.h"
 #include "base/containers/span.h"
 #include "base/containers/span_writer.h"
-#include "brave/components/brave_wallet/common/btc_like_serializer_stream.h"
+#include "brave/components/brave_wallet/browser/zcash/v5_zcash_serializer.h"
+#include "brave/components/brave_wallet/browser/zcash/v6_zcash_serializer.h"
 
 namespace brave_wallet {
 
@@ -22,64 +23,58 @@ namespace {
 
 // https://zips.z.cash/zip-0244
 constexpr char kTransparentHashPersonalizer[] = "ZTxIdTranspaHash";
-constexpr char kSaplingHashPersonalizer[] = "ZTxIdSaplingHash";
-constexpr char kOrchardHashPersonalizer[] = "ZTxIdOrchardHash";
 
 // https://zips.z.cash/zip-0244#txid-digest-1
 constexpr char kTxHashPersonalizerPrefix[] = "ZcashTxHash_";
 
-constexpr uint32_t kV5TxVersion = 5 | 1 << 31 /* overwintered bit */;
-// https://zips.z.cash/protocol/protocol.pdf#txnconsensus
-constexpr uint32_t kV5VersionGroupId = 0x26A7270A;
+}  // namespace
 
-std::array<uint8_t, kZCashDigestSize> Blake2b256(
+// static
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::Blake2b256(
     base::span<const uint8_t> payload,
     base::span<const uint8_t, kBlake2bPersonalizerLength> personalizer) {
   return Blake2bHash<kZCashDigestSize>({payload}, personalizer);
 }
 
-void PushHeader(const ZCashTransaction& tx, BtcLikeSerializerStream& stream) {
-  stream.Push32(kV5TxVersion);
-  stream.Push32(kV5VersionGroupId);
-  stream.Push32(tx.consensus_brach_id());
-  stream.Push32(tx.locktime());
-  stream.Push32(tx.expiry_height());
-}
-
-void PushOutpoint(const ZCashTransaction::Outpoint& outpoint,
-                  BtcLikeSerializerStream& stream) {
+// static
+void ZCashSerializer::PushOutpoint(const ZCashTransaction::Outpoint& outpoint,
+                                   BtcLikeSerializerStream& stream) {
   stream.PushBytes(outpoint.txid);
   stream.Push32(outpoint.index);
 }
 
-void PushOutput(const ZCashTransaction::TxOutput& output,
-                BtcLikeSerializerStream& stream) {
+// static
+void ZCashSerializer::PushOutput(const ZCashTransaction::TxOutput& output,
+                                 BtcLikeSerializerStream& stream) {
   stream.Push64(output.amount);
   stream.PushSizeAndBytes(output.script_pubkey);
 }
 
-// https://zips.z.cash/zip-0244#s-2c-amounts-sig-digest
-std::array<uint8_t, 32> HashAmounts(const ZCashTransaction& tx) {
-  BtcLikeSerializerStream stream;
+// static
+void ZCashSerializer::SerializeTransparentInputs(
+    const ZCashTransaction& tx,
+    BtcLikeSerializerStream& stream) {
+  stream.PushCompactSize(tx.transparent_part().inputs.size());
   for (const auto& input : tx.transparent_part().inputs) {
-    stream.Push64(input.utxo_value);
+    PushOutpoint(input.utxo_outpoint, stream);
+    stream.PushSizeAndBytes(input.script_sig);
+    stream.Push32(input.n_sequence);
   }
-  return Blake2b256(stream.data(),
-                    base::byte_span_from_cstring("ZTxTrAmountsHash"));
 }
 
-// https://zips.z.cash/zip-0244#s-2d-scriptpubkeys-sig-digest
-std::array<uint8_t, 32> HashScriptPubKeys(const ZCashTransaction& tx) {
-  BtcLikeSerializerStream stream;
-  for (const auto& input : tx.transparent_part().inputs) {
-    stream.PushSizeAndBytes(input.script_pub_key);
+// static
+void ZCashSerializer::SerializeTransparentOutputs(
+    const ZCashTransaction& tx,
+    BtcLikeSerializerStream& stream) {
+  stream.PushCompactSize(tx.transparent_part().outputs.size());
+  for (const auto& output : tx.transparent_part().outputs) {
+    PushOutput(output, stream);
   }
-  return Blake2b256(stream.data(),
-                    base::byte_span_from_cstring("ZTxTrScriptsHash"));
 }
 
-std::array<uint8_t, kBlake2bPersonalizerLength> GetHashPersonalizer(
-    const ZCashTransaction& tx) {
+// static
+std::array<uint8_t, kBlake2bPersonalizerLength>
+ZCashSerializer::GetHashPersonalizer(const ZCashTransaction& tx) {
   std::array<uint8_t, kBlake2bPersonalizerLength> result;
   auto span_writer = base::SpanWriter(base::span(result));
   span_writer.Write(base::byte_span_from_cstring(kTxHashPersonalizerPrefix));
@@ -87,8 +82,6 @@ std::array<uint8_t, kBlake2bPersonalizerLength> GetHashPersonalizer(
   DCHECK_EQ(span_writer.remaining(), 0u);
   return result;
 }
-
-}  // namespace
 
 // static
 void ZCashSerializer::SerializeSignature(
@@ -112,59 +105,35 @@ std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashTxIn(
   if (tx_in) {
     PushOutpoint(tx_in->utxo_outpoint, stream);
     stream.Push64(tx_in->utxo_value);
-
     stream.PushSizeAndBytes(tx_in->script_pub_key);
-
     stream.Push32(tx_in->n_sequence);
   }
-
   return Blake2b256(std::move(stream).data(),
                     base::byte_span_from_cstring("Zcash___TxInHash"));
 }
 
 // static
-bool ZCashSerializer::SignTransparentPart(KeyringService& keyring_service,
-                                          const mojom::AccountIdPtr& account_id,
-                                          ZCashTransaction& tx) {
-  auto addresses = keyring_service.GetZCashAddresses(account_id);
-  if (!addresses || addresses->empty()) {
-    return false;
+// https://zips.z.cash/zip-0244#s-2c-amounts-sig-digest
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashAmounts(
+    const ZCashTransaction& tx) {
+  BtcLikeSerializerStream stream;
+  for (const auto& input : tx.transparent_part().inputs) {
+    stream.Push64(input.utxo_value);
   }
+  return Blake2b256(stream.data(),
+                    base::byte_span_from_cstring("ZTxTrAmountsHash"));
+}
 
-  std::map<std::string, mojom::ZCashKeyIdPtr> address_map;
-  for (auto& addr : *addresses) {
-    address_map.emplace(std::move(addr->address_string),
-                        std::move(addr->key_id));
+// static
+// https://zips.z.cash/zip-0244#s-2d-scriptpubkeys-sig-digest
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashScriptPubKeys(
+    const ZCashTransaction& tx) {
+  BtcLikeSerializerStream stream;
+  for (const auto& input : tx.transparent_part().inputs) {
+    stream.PushSizeAndBytes(input.script_pub_key);
   }
-
-  for (size_t input_index = 0;
-       input_index < tx.transparent_part().inputs.size(); ++input_index) {
-    auto& input = tx.transparent_part().inputs[input_index];
-
-    if (!address_map.contains(input.utxo_address)) {
-      return false;
-    }
-
-    auto& key_id = address_map.at(input.utxo_address);
-
-    auto pubkey = keyring_service.GetZCashPubKey(account_id, key_id);
-    if (!pubkey) {
-      return false;
-    }
-
-    auto signature_digest = CalculateSignatureDigest(tx, input);
-
-    auto signature = keyring_service.SignMessageByZCashKeyring(
-        account_id, key_id, base::span(signature_digest));
-
-    if (!signature) {
-      return false;
-    }
-
-    SerializeSignature(tx, input, pubkey.value(), signature.value());
-  }
-
-  return true;
+  return Blake2b256(stream.data(),
+                    base::byte_span_from_cstring("ZTxTrScriptsHash"));
 }
 
 // static
@@ -192,7 +161,7 @@ std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashSequences(
 }
 
 // static
-// https://zips.z.cash/zip-0244#t-2a-prevouts-digest
+// https://zips.z.cash/zip-0244#t-2c-outputs-digest
 std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashOutputs(
     const ZCashTransaction& tx) {
   BtcLikeSerializerStream stream;
@@ -204,175 +173,130 @@ std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashOutputs(
 }
 
 // static
-// https://zips.z.cash/zip-0244#t-1-header-digest
-std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashHeader(
+// https://zips.z.cash/zip-0244 T.2 — transparent txid digest
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashTransparentTxId(
     const ZCashTransaction& tx) {
   BtcLikeSerializerStream stream;
-  PushHeader(tx, stream);
+  if (!tx.transparent_part().inputs.empty() ||
+      !tx.transparent_part().outputs.empty()) {
+    stream.PushBytes(HashPrevouts(tx));
+    stream.PushBytes(HashSequences(tx));
+    stream.PushBytes(HashOutputs(tx));
+  }
   return Blake2b256(stream.data(),
-                    base::byte_span_from_cstring("ZTxIdHeadersHash"));
+                    base::byte_span_from_cstring(kTransparentHashPersonalizer));
 }
 
 // static
-// https://zips.z.cash/zip-0244#txid-digest
-std::array<uint8_t, kZCashDigestSize> ZCashSerializer::CalculateTxIdDigest(
-    const ZCashTransaction& zcash_transaction) {
-  std::array<uint8_t, kZCashDigestSize> header_hash =
-      HashHeader(zcash_transaction);
+// https://zips.z.cash/zip-0244 T.1 — transparent signature digest
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::HashTransparentSignature(
+    const ZCashTransaction& tx,
+    const std::optional<ZCashTransaction::TxInput>& input) {
+  BtcLikeSerializerStream stream;
+  if (!tx.transparent_part().inputs.empty()) {
+    stream.Push8(tx.sighash_type());
+    stream.PushBytes(HashPrevouts(tx));
+    stream.PushBytes(HashAmounts(tx));
+    stream.PushBytes(HashScriptPubKeys(tx));
+    stream.PushBytes(HashSequences(tx));
+    stream.PushBytes(HashOutputs(tx));
+    stream.PushBytes(HashTxIn(input));
+  } else if (!tx.transparent_part().outputs.empty()) {
+    stream.PushBytes(HashPrevouts(tx));
+    stream.PushBytes(HashSequences(tx));
+    stream.PushBytes(HashOutputs(tx));
+  }
+  return Blake2b256(stream.data(),
+                    base::byte_span_from_cstring(kTransparentHashPersonalizer));
+}
 
-  std::array<uint8_t, kZCashDigestSize> transparent_hash;
-  {
-    BtcLikeSerializerStream stream;
+// static
+bool ZCashSerializer::SignTransparentPart(
+    KeyringService& keyring_service,
+    const mojom::AccountIdPtr& account_id,
+    ZCashTransaction& tx,
+    base::FunctionRef<std::array<uint8_t, kZCashDigestSize>(
+        const ZCashTransaction&,
+        const ZCashTransaction::TxInput&)> compute_sig_digest) {
+  auto addresses = keyring_service.GetZCashAddresses(account_id);
+  if (!addresses || addresses->empty()) {
+    return false;
+  }
 
-    // For txid digest (T.2), handle transparent components based on presence.
-    if (!zcash_transaction.transparent_part().inputs.empty() ||
-        !zcash_transaction.transparent_part().outputs.empty()) {
-      // Case: Has transparent inputs or outputs - include all components.
-      stream.PushBytes(HashPrevouts(zcash_transaction));
-      stream.PushBytes(HashSequences(zcash_transaction));
-      stream.PushBytes(HashOutputs(zcash_transaction));
+  std::map<std::string, mojom::ZCashKeyIdPtr> address_map;
+  for (auto& addr : *addresses) {
+    address_map.emplace(std::move(addr->address_string),
+                        std::move(addr->key_id));
+  }
+
+  for (size_t input_index = 0;
+       input_index < tx.transparent_part().inputs.size(); ++input_index) {
+    auto& input = tx.transparent_part().inputs[input_index];
+
+    if (!address_map.contains(input.utxo_address)) {
+      return false;
     }
 
-    // Case: No transparent components - stream remains empty
-    // (BLAKE2b-256("ZTxIdTranspaHash", []))
-    transparent_hash =
-        Blake2b256(stream.data(),
-                   base::byte_span_from_cstring(kTransparentHashPersonalizer));
+    auto& key_id = address_map.at(input.utxo_address);
+
+    auto pubkey = keyring_service.GetZCashPubKey(account_id, key_id);
+    if (!pubkey) {
+      return false;
+    }
+
+    auto signature_digest = compute_sig_digest(tx, input);
+
+    auto signature = keyring_service.SignMessageByZCashKeyring(
+        account_id, key_id, base::span(signature_digest));
+
+    if (!signature) {
+      return false;
+    }
+
+    SerializeSignature(tx, input, pubkey.value(), signature.value());
   }
 
-  std::array<uint8_t, kZCashDigestSize> sapling_hash;
-  {
-    sapling_hash =
-        Blake2b256({}, base::byte_span_from_cstring(kSaplingHashPersonalizer));
-  }
+  return true;
+}
 
-  std::array<uint8_t, kZCashDigestSize> orchard_hash;
-  {
-    orchard_hash = zcash_transaction.orchard_part().digest.value_or(
-        Blake2b256({}, base::byte_span_from_cstring(kOrchardHashPersonalizer)));
-  }
+// Dispatch entry points.
 
-  std::array<uint8_t, kZCashDigestSize> digest_hash;
-  {
-    BtcLikeSerializerStream stream;
-    stream.PushBytes(header_hash);
-    stream.PushBytes(transparent_hash);
-    stream.PushBytes(sapling_hash);
-    stream.PushBytes(orchard_hash);
-
-    digest_hash =
-        Blake2b256(stream.data(), GetHashPersonalizer(zcash_transaction));
-  }
-
-  std::reverse(digest_hash.begin(), digest_hash.end());
-
-  return digest_hash;
+// static
+std::array<uint8_t, kZCashDigestSize> ZCashSerializer::CalculateTxIdDigest(
+    const ZCashTransaction& zcash_transaction) {
+  return zcash_transaction.v6_part()
+             ? ZCashV6Serializer::CalculateTxIdDigest(zcash_transaction)
+             : ZCashV5Serializer::CalculateTxIdDigest(zcash_transaction);
 }
 
 // static
-// https://zips.z.cash/zip-0244#signature-digest
 std::array<uint8_t, kZCashDigestSize> ZCashSerializer::CalculateSignatureDigest(
     const ZCashTransaction& zcash_transaction,
     const std::optional<ZCashTransaction::TxInput>& input) {
-  std::array<uint8_t, kZCashDigestSize> header_hash =
-      HashHeader(zcash_transaction);
-
-  std::array<uint8_t, kZCashDigestSize> transparent_hash;
-  {
-    BtcLikeSerializerStream stream;
-
-    if (!zcash_transaction.transparent_part().inputs.empty()) {
-      stream.Push8(zcash_transaction.sighash_type());
-      stream.PushBytes(HashPrevouts(zcash_transaction));
-      stream.PushBytes(HashAmounts(zcash_transaction));
-      stream.PushBytes(HashScriptPubKeys(zcash_transaction));
-      stream.PushBytes(HashSequences(zcash_transaction));
-      stream.PushBytes(HashOutputs(zcash_transaction));
-      stream.PushBytes(HashTxIn(input));
-    } else if (!zcash_transaction.transparent_part().outputs.empty()) {
-      // If we are producing a hash for either a coinbase transaction,
-      // or a non-coinbase transaction that has no transparent inputs,
-      // the value of transparent_sig_digest is identical to the value
-      // specified in section T.2.
-      stream.PushBytes(HashPrevouts(zcash_transaction));
-      stream.PushBytes(HashSequences(zcash_transaction));
-      stream.PushBytes(HashOutputs(zcash_transaction));
-    }
-
-    // Case: No transparent components - stream remains empty.
-    transparent_hash =
-        Blake2b256(stream.data(),
-                   base::byte_span_from_cstring(kTransparentHashPersonalizer));
-  }
-
-  std::array<uint8_t, kZCashDigestSize> sapling_hash;
-  {
-    sapling_hash =
-        Blake2b256({}, base::byte_span_from_cstring(kSaplingHashPersonalizer));
-  }
-
-  std::array<uint8_t, kZCashDigestSize> orchard_hash;
-  {
-    orchard_hash = zcash_transaction.orchard_part().digest.value_or(
-        Blake2b256({}, base::byte_span_from_cstring(kOrchardHashPersonalizer)));
-  }
-
-  std::array<uint8_t, kZCashDigestSize> digest_hash;
-  {
-    BtcLikeSerializerStream stream;
-    stream.PushBytes(header_hash);
-    stream.PushBytes(transparent_hash);
-    stream.PushBytes(sapling_hash);
-    stream.PushBytes(orchard_hash);
-
-    digest_hash =
-        Blake2b256(stream.data(), GetHashPersonalizer(zcash_transaction));
-  }
-
-  return digest_hash;
+  return zcash_transaction.v6_part()
+             ? ZCashV6Serializer::CalculateSignatureDigest(zcash_transaction,
+                                                           input)
+             : ZCashV5Serializer::CalculateSignatureDigest(zcash_transaction,
+                                                           input);
 }
 
 // static
-// https://zips.z.cash/zip-0225
 std::vector<uint8_t> ZCashSerializer::SerializeRawTransaction(
     const ZCashTransaction& zcash_transaction) {
-  BtcLikeSerializerStream stream;
+  return zcash_transaction.v6_part()
+             ? ZCashV6Serializer::SerializeRawTransaction(zcash_transaction)
+             : ZCashV5Serializer::SerializeRawTransaction(zcash_transaction);
+}
 
-  PushHeader(zcash_transaction, stream);
-
-  // Tx In
-  {
-    // Inputs size
-    stream.PushCompactSize(zcash_transaction.transparent_part().inputs.size());
-    for (const auto& input : zcash_transaction.transparent_part().inputs) {
-      // Outpoint
-      PushOutpoint(input.utxo_outpoint, stream);
-      stream.PushSizeAndBytes(input.script_sig);
-      // Sequence
-      stream.Push32(input.n_sequence);
-    }
-  }
-
-  // Tx Out
-
-  // Outputs size
-  stream.PushCompactSize(zcash_transaction.transparent_part().outputs.size());
-  for (const auto& output : zcash_transaction.transparent_part().outputs) {
-    PushOutput(output, stream);
-  }
-
-  // Sapling
-  stream.PushCompactSize(0u);
-  stream.PushCompactSize(0u);
-
-  // Orchard
-  if (zcash_transaction.orchard_part().raw_tx) {
-    stream.PushBytes(zcash_transaction.orchard_part().raw_tx.value());
-  } else {
-    stream.PushCompactSize(uint8_t{0});
-  }
-
-  return std::move(stream).Take();
+// static
+bool ZCashSerializer::SignTransparentPart(KeyringService& keyring_service,
+                                          const mojom::AccountIdPtr& account_id,
+                                          ZCashTransaction& tx) {
+  return tx.v6_part()
+             ? ZCashV6Serializer::SignTransparentPart(keyring_service,
+                                                      account_id, tx)
+             : ZCashV5Serializer::SignTransparentPart(keyring_service,
+                                                      account_id, tx);
 }
 
 }  // namespace brave_wallet
